refactor(crowdfundngs): Replace sendPaymentSuccessful with lib/Mail
